chore: rename classic cad asset mapping cache file
